# Conflicts:
#	src/README.md
#	src/apisvr/http/things.api
#	src/apisvr/internal/config/config.go
#	src/apisvr/internal/handler/routes.go
#	src/apisvr/internal/logic/assemble.go
#	src/apisvr/internal/svc/serviceContext.go
#	src/apisvr/internal/types/types.go
#	src/dmsvr/pb/dm/dm.pb.go
#	src/syssvr/pb/sys/sys.pb.go

+++ src/dmsvr/internal/logic/devicegroup/assemble.go
@@ -16,7 +16,7 @@ func ToGroupInfoPb(ro *relationDB.DmGroupInfo) *dm.GroupInfo {
 	return &dm.GroupInfo{
 		GroupID:     ro.GroupID,
 		ParentID:    ro.ParentID,
-		ProjectID:   ro.ProjectID,
+		ProjectID:   int64(ro.ProjectID),
 		ProductName: productName,
 		GroupName:   ro.GroupName,
 		ProductID:   ro.ProductID,
